+++ packages/nimble-components/src/theme-provider/design-tokens.ts
@@ -82,8 +82,15 @@ import {
     GroupLabel1LineHeight,
     ControlLabel1LineHeight,
     ButtonLabel1LineHeight,
-    TooltipCaptionLineHeight
+    TooltipCaptionLineHeight,
+    Information100LightUi,
+    Information100DarkUi
 } from '@ni/nimble-tokens/dist/styledictionary/js/tokens';
+import {
+    modalBackdropColorThemeColor,
+    modalBackdropColorThemeDark,
+    modalBackdropColorThemeLight
+} from './design-tokens-static';
 import { Theme } from './types';
 import { tokenNames, styleNameFromTokenName } from './design-token-names';
 import { theme } from '.';
@@ -150,6 +157,10 @@ export const passColor = DesignToken.create<string>(
     styleNameFromTokenName(tokenNames.passColor)
 ).withDefault((element: HTMLElement) => getPassColorForTheme(element));
 
+export const informationColor = DesignToken.create<string>(
+    styleNameFromTokenName(tokenNames.informationColor)
+).withDefault((element: HTMLElement) => getInformationColorForTheme(element));
+
 export const borderHoverColor = DesignToken.create<string>(
     styleNameFromTokenName(tokenNames.borderHoverColor)
 ).withDefault((element: HTMLElement) => getColorForTheme(element, DigitalGreenLight, DigitalGreenLight, White));
@@ -158,6 +169,10 @@ export const borderHoverColor = DesignToken.create<string>(
 export const iconColor = DesignToken.create<string>(
     styleNameFromTokenName(tokenNames.iconColor)
 ).withDefault((element: HTMLElement) => getColorForTheme(element, Black91, Black15, White));
+
+export const modalBackdropColor = DesignToken.create<string>(
+    styleNameFromTokenName(tokenNames.modalBackdropColor)
+).withDefault((element: HTMLElement) => getModalBackdropForTheme(element));
 
 export const popupBoxShadowColor = DesignToken.create<string>(
     styleNameFromTokenName(tokenNames.popupBoxShadowColor)
@@ -658,6 +673,15 @@ function getPassColorForTheme(element: HTMLElement): string {
     return getColorForTheme(element, Pass100LightUi, Pass100DarkUi, White);
 }
 
+function getInformationColorForTheme(element: HTMLElement): string {
+    return getColorForTheme(
+        element,
+        Information100LightUi,
+        Information100DarkUi,
+        White
+    );
+}
+
 function getDefaultLineColorForTheme(element: HTMLElement): string {
     return getColorForTheme(element, Black91, Black15, White);
 }
@@ -681,4 +705,17 @@ function getFillHoverColorForTheme(element: HTMLElement): string {
 
 function getFillDownColorForTheme(element: HTMLElement): string {
     return getColorForTheme(element, Black91, Black15, White);
+}
+
+function getModalBackdropForTheme(element: HTMLElement): string {
+    switch (theme.getValueFor(element)) {
+        case Theme.light:
+            return modalBackdropColorThemeLight;
+        case Theme.dark:
+            return modalBackdropColorThemeDark;
+        case Theme.color:
+            return modalBackdropColorThemeColor;
+        default:
+            return modalBackdropColorThemeLight;
+    }
 }